Rewrote interpreter to first compile source code to optimized form
Instead of executing the source code verbatim, generate an intermediate
representation with less (but more powerful) commands. The assumption
was that this would improve execution times, but alas it seems that
things like factorial.bf became about 25% slower.

However, the code looks nicer to me so let's keep this in a branch and
see whether we can optimize later.