Added new 'zeroCells' optimizer
This optimizer can introduce a new dedicated 'SetCell' instruction to
the program, which sets the current cell to a specific value. This is
useful for constructs like [-] which clearly just sets the current cell
to zero. The optimizer relies on the 'merge' phase being done already
so that we can handle cases like [-++--+-] for free.