Enable building with latest GHC (and Stackage LTS 10.3)
This adjusts the .cabal file such that the library constraints work with
the latest GHC and base libraries.

I also cleaned things up a bit such that not just a single executable is
built, but instead a library is compiled plus an executable using that
library. When doing so, I shuffled the source code around a bit:

This seems to simplify building a library and an executable in a single
.cabal file: by having separate directory, we avoid that the library
source code gets re-built when building the executable (since both have
their source code below 'src').